test(platform-ui): move action-flows vitest tests to src/test/integration/action-flows/
Per the 5-level pyramid, these are L2 integration tests (vitest +
mocked services + local Node echo server), not Playwright e2e.
Their location in e2e/ was an artefact of isolating their boot
footprint via a dedicated vitest config. Move them to the L2 home
and keep them excluded from default `pnpm test:run`.
