+++ packages/platform-ui/vitest.config.ts
@@ -25,6 +25,6 @@ export default defineConfig({
     environment: 'jsdom',
     setupFiles: ['./src/test/setup.ts'],
     globals: true,
-    exclude: ['e2e/**', 'node_modules/**'],
+    exclude: ['e2e/**', 'node_modules/**', 'src/test/integration/action-flows/**'],
   },
 });
